add some code to commands
