+++ bot/utils.py
@@ -6,3 +6,4 @@ from aiogram.fsm.state import State, StatesGroup
 class Route(StatesGroup):
     route = State()
     exhibit = State()
+    review = State()
