+++ requirements.txt
@@ -1,6 +1,5 @@
 aiofiles==23.1.0
 aiogram==3.0.0rc1
-aiogram-dialog==1.9.0
 aiohttp==3.8.5
 aiosignal==1.3.1
 annotated-types==0.5.0
